pyproject.toml
[project]
name = "palo-alto-grafana-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.0",
    "colorama>=0.4.6",
    "pandas>=2.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
    "urllib3>=2.0.0",
    "xmltodict>=0.13.0",
]
